requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
beautifulsoup4==4.13.3
    # via documentation (pyproject.toml)
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
click==8.1.8
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via click
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via
    #   documentation (pyproject.toml)
    #   mcp
httpx-sse==0.4.0
    # via mcp
idna==3.10
    # via
    #   anyio
    #   httpx
markdown-it-py==3.0.0
    # via rich
mcp==1.5.0
    # via documentation (pyproject.toml)
mdurl==0.1.2
    # via markdown-it-py
pydantic==2.10.6
    # via
    #   mcp
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.8.1
    # via mcp
pygments==2.19.1
    # via rich
python-dotenv==1.0.1
    # via
    #   mcp
    #   pydantic-settings
rich==13.9.4
    # via typer
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via anyio
soupsieve==2.6
    # via beautifulsoup4
sse-starlette==2.2.1
    # via mcp
starlette==0.46.1
    # via
    #   mcp
    #   sse-starlette
typer==0.15.2
    # via mcp
typing-extensions==4.12.2
    # via
    #   anyio
    #   beautifulsoup4
    #   pydantic
    #   pydantic-core
    #   typer
uvicorn==0.34.0
    # via mcp
